+++ src/sns/report/live/types/mod.rs
@@ -7,7 +7,6 @@
 mod deployed;
 mod neurons;
 mod proposals;
-mod token;
 
 pub(in crate::sns::report::live) use deployed::{
     DeployedSns, GetMetadataRequest, GetMetadataResponse, ListDeployedSnsesRequest,
@@ -20,8 +19,4 @@ pub(in crate::sns::report::live) use proposals::{
     GetProposalRequest, GetProposalResponse, GetProposalResult, ListProposalsRequest,
     ListProposalsResponse, SnsGovernanceBallot, SnsGovernanceProposalData, SnsProposalId, SnsTopic,
     SnsTopicSelector,
-};
-pub(in crate::sns::report) use token::IcrcMetadataValue;
-pub(in crate::sns::report::live) use token::{
-    GetIndexPrincipalError, GetIndexPrincipalResult, IcrcAccount, IcrcSupportedStandard,
 };